# rust: Use cmac and pmac crates from RustCrypto