These crates provide faster implementations of the CMAC and PMAC constructions.

Before:

```
test bench::bench_aes_pmac_siv_128_encrypt_1024_bytes  ... bench:         964 ns/iter (+/- 185) = 1062 MB/s
test bench::bench_aes_pmac_siv_128_encrypt_128_bytes   ... bench:         327 ns/iter (+/- 60) = 391 MB/s
test bench::bench_aes_pmac_siv_128_encrypt_16384_bytes ... bench:      11,384 ns/iter (+/- 2,166) = 1439 MB/s
test bench::bench_aes_siv_128_encrypt_1024_bytes       ... bench:       1,275 ns/iter (+/- 215) = 803 MB/s
test bench::bench_aes_siv_128_encrypt_128_bytes        ... bench:         261 ns/iter (+/- 72) = 490 MB/s
test bench::bench_aes_siv_128_encrypt_16384_bytes      ... bench:      18,235 ns/iter (+/- 3,579) = 898 MB/s
```

After:

```
test bench::bench_aes_pmac_siv_128_encrypt_1024_bytes  ... bench:         771 ns/iter (+/- 112) = 1328 MB/s
test bench::bench_aes_pmac_siv_128_encrypt_128_bytes   ... bench:         231 ns/iter (+/- 40) = 554 MB/s
test bench::bench_aes_pmac_siv_128_encrypt_16384_bytes ... bench:      10,078 ns/iter (+/- 1,890) = 1625 MB/s
test bench::bench_aes_siv_128_encrypt_1024_bytes       ... bench:       1,190 ns/iter (+/- 430) = 860 MB/s
test bench::bench_aes_siv_128_encrypt_128_bytes        ... bench:         289 ns/iter (+/- 70) = 442 MB/s
test bench::bench_aes_siv_128_encrypt_16384_bytes      ... bench:      17,169 ns/iter (+/- 3,445) = 954 MB/s
```